Fix pipeline

+++ run-tests.sh
@@ -10,6 +10,7 @@ cmake -S . -B build -G Ninja
 
 git diff --name-only HEAD~9 HEAD | grep -E 'Leetcode/.*cpp' | while read -r file_path
 do
+    [ -f $file_path ] || continue
     printf "\n-------- Running tests for file %s --------\n" "$file_path"
     cp "$file_path" ./leetcode.cpp
     cmake --build build
